Universe: tick and setRules

+++ C++-Labs/lab_2_game_of_life/src/core/Field.cpp
@@ -10,13 +10,13 @@ std::vector<bool> &Field::operator[](size_t x) {
     return field_data_[x];
 }
 
-uint32_t Field::getNeighboursCount(size_t x, size_t y) const {
+uint8_t Field::getNeighboursCount(size_t x, size_t y) const {
     size_t leftColumn = (x + this->width() - 1) % this->width();
     size_t rightColumn = (x + 1) % this->width();
     size_t topRow = (y + this->height() - 1) % this->height();
     size_t bottomRow = (y + 1) % this->height();
 
-    uint32_t neighboursCount = 0;
+    uint8_t neighboursCount = 0;
     neighboursCount += field_data_[leftColumn][topRow];
     neighboursCount += field_data_[leftColumn][y];
     neighboursCount += field_data_[leftColumn][bottomRow];
